db파일 exsist체크 추가

+++ src/mysql_to_sqlite3/__main__.py
@@ -5,7 +5,8 @@ class Main:
     def __init__(self):
         self.path = "../../Assets/StreamingAssets/Reference_Data.db"
 
-        os.remove(self.path)
+        if (os.path.exists(self.path)):
+            os.remove(self.path)
 
         converter = MySQLtoSQLite(
             sqlite_file=self.path,
